requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-restx==1.3.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-flask==1.3.0
borneo==5.4.2
oci==2.125.3
gunicorn==21.2.0
PyJWT
cryptography==41.0.7
pymongo==4.6.0
dnspython==2.4.2
